// client/src/generated/api.ts
export interface UserBody {
  /** @example John */
  firstName: string;

  /** @example Doe */
  lastName: string;

  /**
   * @format email
   * @example john.doe@foo.bar
   */
  email: string;

  /**
   * @format password
   * @example password
   */
  password: string;
}

export type UserFullBody = UserBody & { id: string };

export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, 'body' | 'bodyUsed'>;

export interface FullRequestParams extends Omit<RequestInit, 'body'> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: ResponseFormat;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<
  FullRequestParams,
  'body' | 'method' | 'query' | 'path'
>;
// ...
export enum ContentType {
  Json = 'application/json',
  FormData = 'multipart/form-data',
  UrlEncoded = 'application/x-www-form-urlencoded',
}
// ...
export class Api<
  SecurityDataType extends unknown
> extends HttpClient<SecurityDataType> {
  user = {
    /**
     * @description Fetch all users
     *
     * @tags Users
     * @name UserList
     * @request GET:/user
     */
    userList: (params: RequestParams = {}) =>
      this.request<UserFullBody[], any>({
        path: `/user`,
        method: 'GET',
        format: 'json',
        ...params,
      }),

    /**
     * @description Create a new user
     *
     * @tags Users
     * @name UserCreate
     * @request POST:/user
     */
    userCreate: (user: UserBody, params: RequestParams = {}) =>
      this.request<UserFullBody, any>({
        path: `/user`,
        method: 'POST',
        body: user,
        format: 'json',
        ...params,
      }),

    /**
     * No description
     *
     * @tags Users
     * @name UserDetail
     * @request GET:/user/{id}
     */
    userDetail: (id: string, params: RequestParams = {}) =>
      this.request<UserFullBody, void>({
        path: `/user/${id}`,
        method: 'GET',
        format: 'json',
        ...params,
      }),

    /**
     * No description
     *
     * @tags Users
     * @name UserUpdate
     * @request PUT:/user/{id}
     */
    userUpdate: (id: string, user: UserBody, params: RequestParams = {}) =>
      this.request<UserFullBody, void>({
        path: `/user/${id}`,
        method: 'PUT',
        body: user,
        format: 'json',
        ...params,
      }),

    /**
     * No description
     *
     * @tags Users
     * @name UserDelete
     * @request DELETE:/user/{id}
     */
    userDelete: (id: string, params: RequestParams = {}) =>
      this.request<boolean, any>({
        path: `/user/${id}`,
        method: 'DELETE',
        format: 'json',
        ...params,
      }),
  };
  spec = {
    /**
     * No description
     *
     * @tags Specification
     * @name SpecList
     * @request GET:/spec
     */
    specList: (params: RequestParams = {}) =>
      this.request<void, any>({
        path: `/spec`,
        method: 'GET',
        ...params,
      }),
  };
}